Fixed recursive call in Loader
`setTimeOut` requires you to pass a function. `parseFrame` does not return a function. It (apparently) returns something string-like. That's also a constructor for `setTimeOut` but then one that runs `eval` on the string. Nobody wants that, and that's also the reason modern browsers block it. I think the intended behaviour was to call `setTimeOut` with a function, so `parseFrame` is called after a timeout and recursively calls itself. 

I didn't know how to test the code, so I'll leave that up to you